定数バッファを修正・定数バッファデータクラス作成・DirectX11用とわかるようにファイル名変更・必要のない情報を保持しないように修正・デバッグ描画で定数バッファを作成するように修正
定数バッファ作成をリファレンスから作成しないように修正
定数バッファマネージャーをシェーダ―マネージャーに渡さないように修正
定数バッファデータクラスを作成（バッファ作成・保持を行う）
シェーダーに名前やエントリーポイント情報が不必要なため削除
デバッグ描画で定数バッファを作成するように修正

+++ 3DAction_Project/3DAction_Project/VertexBufferIData.cpp
@@ -16,7 +16,7 @@ bool VertexBufferData::CreateVertexBuffer(
 	const void* vertices,
 	int vertexCount,
 	int vertexMaxCount,
-	int stride,
+	size_t stride,
 	D3D11_PRIMITIVE_TOPOLOGY primitiveType,
 	D3D11_USAGE usage,
 	D3D11_CPU_ACCESS_FLAG flag)
